+++ packages/android_alarm_manager/android/src/main/java/io/flutter/plugins/androidalarmmanager/AlarmService.java
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package io.flutter.androidalarmmanager;
+package io.flutter.plugins.androidalarmmanager;
 
 import android.app.Activity;
 import android.app.AlarmManager;