+++ packages/firebase_admob/ios/Classes/FirebaseAdmobPlugin.h
@@ -4,5 +4,8 @@
 
 #import <Flutter/Flutter.h>
 
+#define FLTLogWarning(format, ...) \
+  NSLog((@"FirebaseAdMobPlugin <warning> " format), ##__VA_ARGS__)
+
 @interface FLTFirebaseAdMobPlugin : NSObject<FlutterPlugin>
 @end