+++ packages/firebase_auth/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.4.3
+
+* Moved to the io.flutter.plugins organization.
+
+## 0.4.2
+
+* Added support for changing user data
+
 ## 0.4.1
 
 * Simplified and upgraded Android project template to Android SDK 27.