+++ packages/firebase_database/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.4
+
+* Allow null values for Query startAt, endAt, and equalTo
+
+## 0.3.3
+
+* Support to specify a database by URL if required 
+
 ## 0.3.2
 
 * Fix warnings from the Dart 2.0 analyzer.